ProjectionPursuit

Documentation for ProjectionPursuit.

ProjectionPursuit

Functions

Index

Functions

ProjectionPursuit.gensphere — Method

gensphere(N::Int64, d::Int64)

Generate low discrepancy sequences of length N on a d dimension unit sphere.

Roughly speaking, a low discrepancy sequence is a sequence of samples of a random variable has the property that the number of points within a region is proportional to the size of the region.

This function returns a N-by-d matrix, where each row represents a sample of a random vector that is uniformly distributed on a d-dimensional unit sphere.

This function is based on Brauchart, Johann S., Josef Dick, and Lou Fang. "Spatial low-discrepancy sequences, spherical cone discrepancy, and applications in financial modeling." Journal of Computational and Applied Mathematics 286 (2015): 28-53.

ProjectionPursuit.fastgensphere — Method

fastgensphere(N::Int64, d::Int64; par::Bool=true)

A faster way to generate lds of length N on a d dimension unit sphere.

Similar to gensphere(). This function is much faster and support parallelism, but may degenarate when d is large.

ProjectionPursuit.fromsphere — Method

fromsphere(theta::Vector{Float64})

Transform a d-dimensional spherical coordinate to the corresponding d+1 Cartesian coordinate.

See also tosphere.

ProjectionPursuit.tosphere — Method

tosphere(s::Vector{Float64})

Transform a d-dimensional Cartesian coordinate to the corresponding d-1 spherical coordinate.

See also fromsphere.

ProjectionPursuit.ProjectionPursuitRes — Type

ProjectionPursuitRes

Returned object of projection_pursuit or sphere_optimize.

Fields

object_vals::Vector{Float64}: projetion pursuit index. Analogous to eigenvalues in PCA.

proj::Matrix{Float64}: projetion pursuit direction. Analogous to eigenvectors in PCA.

Index

ProjectionPursuit.ProjectionPursuitRes

ProjectionPursuit.fastgensphere

ProjectionPursuit.fromsphere

ProjectionPursuit.gensphere

ProjectionPursuit.tosphere